+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Limitation of maximum project drafts.
+- Project view counting.
 
 ## [0.0.5] - 2019-08-09
 ### Added